000_dcm_convert.py

+++ 007_quality_control.py
@@ -43,6 +43,7 @@ def extract_quality_metrics(population, workspace_dir):
         flash_wm  = os.path.join(subject_dir, 'REGISTRATION/FLASH_WM_opt.nii.gz')
         flash_cm  = os.path.join(subject_dir, 'REGISTRATION/FLASH_CSF_opt.nii.gz')
 
+
         def calc_spatial_metrics(img, mask, gm, wm,cm):
 
             # Load data
@@ -68,27 +69,31 @@ def extract_quality_metrics(population, workspace_dir):
 
             return qc_cnr, qc_snr, qc_fwhm[3], qc_efc, qc_fber, qc_qi1
 
-        uni_qc_cnr, uni_qc_snr, uni_qc_fwhm, uni_qc_efc, uni_qc_fber, uni_qc_qi1 = calc_spatial_metrics(mp2rage_uni, mp2rage_mas, mp2rage_gm, mp2rage_wm, mp2rage_cm)
-        mag_qc_cnr, mag_qc_snr, mag_qc_fwhm, mag_qc_efc, mag_qc_fber, mag_qc_qi1 = calc_spatial_metrics(flash_mag, flash_mas, flash_gm, flash_wm, flash_cm)
+        if not os.path.isfile(os.path.join(qc_dir, 'QC.csv')):
 
-        cols = ['SNR_UNI', 'CNR_UNI', 'FBER_UNI', 'EFC_UNI', 'FWHM_UNI', 'QI1_UNI',
-                'SNR_MAG', 'CNR_MAG', 'FBER_MAG', 'EFC_MAG', 'FWHM_MAG', 'QI1_MAG',
-                ]
-        df = pd.DataFrame(columns=cols, index=['%s' % subject])
+            uni_qc_cnr, uni_qc_snr, uni_qc_fwhm, uni_qc_efc, uni_qc_fber, uni_qc_qi1 = calc_spatial_metrics(mp2rage_uni, mp2rage_mas, mp2rage_gm, mp2rage_wm, mp2rage_cm)
+            mag_qc_cnr, mag_qc_snr, mag_qc_fwhm, mag_qc_efc, mag_qc_fber, mag_qc_qi1 = calc_spatial_metrics(flash_mag, flash_mas, flash_gm, flash_wm, flash_cm)
 
-        df.loc[subject]['SNR_UNI']  = uni_qc_snr
-        df.loc[subject]['CNR_UNI']  = uni_qc_cnr
-        df.loc[subject]['FBER_UNI'] = uni_qc_fber
-        df.loc[subject]['EFC_UNI']  = uni_qc_efc
-        df.loc[subject]['FWHM_UNI'] = uni_qc_fwhm
-        df.loc[subject]['QI1_UNI']  = uni_qc_qi1
-        df.loc[subject]['SNR_MAG']  = mag_qc_snr
-        df.loc[subject]['CNR_MAG']  = mag_qc_cnr
-        df.loc[subject]['FBER_MAG'] = mag_qc_fber
-        df.loc[subject]['EFC_MAG']  = mag_qc_efc
-        df.loc[subject]['FWHM_MAG'] = mag_qc_fwhm
-        df.loc[subject]['QI1_MAG']  = mag_qc_qi1
-        df.to_csv(os.path.join(qc_dir, 'QC.csv'))
+            cols = ['SNR_UNI', 'CNR_UNI', 'FBER_UNI', 'EFC_UNI', 'FWHM_UNI', 'QI1_UNI',
+                    'SNR_MAG', 'CNR_MAG', 'FBER_MAG', 'EFC_MAG', 'FWHM_MAG', 'QI1_MAG',
+                    ]
+            df = pd.DataFrame(columns=cols, index=['%s' % subject])
+
+            df.loc[subject]['SNR_UNI']  = uni_qc_snr
+            df.loc[subject]['CNR_UNI']  = uni_qc_cnr
+            df.loc[subject]['FBER_UNI'] = uni_qc_fber
+            df.loc[subject]['EFC_UNI']  = uni_qc_efc
+            df.loc[subject]['FWHM_UNI'] = uni_qc_fwhm
+            df.loc[subject]['QI1_UNI']  = uni_qc_qi1
+            df.loc[subject]['SNR_MAG']  = mag_qc_snr
+            df.loc[subject]['CNR_MAG']  = mag_qc_cnr
+            df.loc[subject]['FBER_MAG'] = mag_qc_fber
+            df.loc[subject]['EFC_MAG']  = mag_qc_efc
+            df.loc[subject]['FWHM_MAG'] = mag_qc_fwhm
+            df.loc[subject]['QI1_MAG']  = mag_qc_qi1
+            df.to_csv(os.path.join(qc_dir, 'QC.csv'))
 
 
-extract_quality_metrics(controls_a + patients_a + lemon_population, workspace_iron)
+extract_quality_metrics(patients_a, workspace_iron)
+extract_quality_metrics(lemon_population, workspace_iron)
+extract_quality_metrics(controls_a, workspace_iron)
